test-appliance: install drgn into the kvm and gce test appliance
drgn[1][2] is a super handy Python based debugging tool for the Linux
kernel.  It can be used to debug crash dumps and for many use cases it
is simpler to examine kernel data structures compared to the crash[3]
tool.  More importantly for our purposes here, it can be used to look
at the data structures for the running kernel, although it does
require that the debug info files be installed.

[1] https://github.com/osandov/drgn
[2] https://developers.facebook.com/blog/post/2021/12/09/drgn-how-linux-kernel-team-meta-debugs-kernel-scale/
[3] https://man7.org/linux/man-pages/man8/crash.8.html

Signed-off-by: Theodore Ts'o <tytso@mit.edu>