Create and reorganzize top-level directories
The directory structure in xfstests-bld reflects its original design
as a hermetic build system for xfstests.  However, these days, its
primary focus is the an test appliance featuring xfstests (but these
days, we also support blktests the Phoronix test suite, etc.) and
drivers to run the test appliance on a variety of environments
(android, gce, and kvm).

So move around a large number of files and directories so that they
are sorted into four top-level directories:

   fstests-bld		The hermetic build system for xfstests
   test-appliance	The test runner infrastructure for xfstests
   			(and blktests and Phoronix Test Suite....)
   run-fstests		The test runner command line utilities, namely
   			kvm-xfstests, gce-xfstests, etc.
   build-kernel		Utilities to build and configure the Linux kernel
   			in a standard way which is easy for the test runners
			to run.

There is a script in fstests-bld/misc/post-reorg-cleanup which may be
helpful in moving the external repos and other files to the proper
places after this reorganization.  Please take a look at it before
running it run it with the --no-action argument first.  Since the most
developers will only need to run the script once, it may be a little
rough, and it does delete some files and directories.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>