test-appliance: add kernel version conditionals using cpp to exclude files
Process exclude files using cpp, so we can conditionalize based on the
kernel version that we are testing.   Example usage:

// The test generic/471 tests the RWF_NOWAIT flag; however
// how it is supposed to work with file systems is disputed,
// and not all device drivers support it.  As a result
// it doesn't work if an LVM device driver is in use...
generic/471

Use these conditionals to the global excludes since these issues are
now solved on modern (>= 5.10 LTS) kernels.

In addition move the above example to the exclude.gce file, since even
on 5.4 and earlier kernels, it's fine to run this test on kvm; it's
only an issue when runnning gce-xfstests.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>