test-appliance: factor out dax group exclusions to the global_exclude file
Teach runtests.sh to #define IS_DAX if the test config is named "dax",
and use this so if IS_DAX_CONFIG is not defined, to exclude all tests
in the dax group.  This allows us to centralize all of the dax
exclusions into the global_exclude file, simplifying (and in many
cases, removing) test-case specific exclude files.

In addition, use the cpp #include statement to simplify the exclude
files for ext4/encrypt_1k and ext4/orphan_file_1k.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>