gce-xfstests: don't update gce_xfstests.config when it's unnecessary
The commands "gsutil stat" and "gsutil hash" take approximately a 1s
and 0.5s, respectively.  So save a copy of what the config is supposed
to be in /tmp/.gce_xfstests.config.$USER, and compare against that
local file if it exists, and skip comparing against the version in
Google Cloud Storage if the config file hasn't changed.

This means that if the user updates ~/.config/gce-xfststs it will
automatically get updated in the GCS bucket; however, if the cached
version of the config file hasn't changed, don't even try to update
the version in GCE.

Also don't update the gce_xfstests.config file in the GCS bucket in
the LTM or KCS servers, since this file is created for the benefit of
the LTM and KCS VM's.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>