test-appliance: force 4 KB block size for bigalloc, bigalloc_inline
The cfg file for the bigalloc test configuration does not explicitly
define the file system block size as is done for the 4k configuration,
although the intent is to test a file system with 4 KB blocks and 64 KB
clusters.  At least one test, shared/298, runs with a block size of
1 KB instead under bigalloc because it creates a file system image less
than 512 MB in size, a result of the mke2fs.conf block size rule
for small files.

shared/298 currently fails when run under bigalloc with 1 KB blocks.
When the block size is set to 4 KB for the test, it passes.

Explicitly defining the bigalloc block size will help avoid similar
surprises in current or future tests written to use small test files.
Make the same change to the bigalloc_inline config file while we're
at it.

v2:  Modify the names of the bigalloc test configurations using 4 KB
block sizes to explicitly reflect the block size.  Change the
documentation and supporting files to reflect this.  Bring the
bigalloc_4k_inline.exclude file up to date (and propagate a change to
the other .exclude files).  Add a new test configuration for bigalloc
with 64k blocks, but don't add this configuration to the default list
of all tests to be run for now.

The bigalloc_64k and huge_bigalloc_4k configurations are untested.  The
huge_bigalloc_4k.exclude file will likely need further work if used.

Signed-off-by: Eric Whitney <enwlinux@gmail.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>