test-appliance: install arm64 cross-compiler in the GCE test appliance
Install the package gcc-aarch6t4-linux-gnu so we can support
cross-compiling the arm64 kernel on an x86 build server.  Also the
default gcc, instead of gcc-9, since for Debian Bullseye (currently
stable as of the time of this writing) the default gcc compiler
version is gcc 10, and that's the same version of gcc installed by
gcc-aarch6t4-linux-gnu.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>