run-fstests: get/infer the architecture from the kernel file
When uploading a file using "gce-xfstests upload-kernel", infer the
architecture from the kernel to be uploaded, and set it in the Google
Cloud Storage metadata.

When running a test, get the architecture from the kernel file, and if
the architecture is explicitly specified, check to make sure the
kernel architecture matches the architecture specified on the command
line.  If the architecture is not explicitly specified, use the
architecture of the kernel as the architecture that should be used for
the test.

The net result of this change is that if an arm64 kernel is uploaded,
for example, using the command "gce-xfstests --arm64 upload-kernel",
the uploaded kernel will be tagged with the appropriate metadata, and
so when a test is launched via "gce-xfstests --kernel
gs://$GS_BUCKET/kernel.deb", the architecture type arm64 will be
automatically inferred, and an ARM64 VM will be used.

Similarly, if kvm-xfstests is passed an explicit kernel --- for
example "kvm-xfstests --kernel /build/ext4-arm64/arch/arm64/boot/Image
shell" qemu-system-aarch64 will be automatically used.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>