gce-xfstests: fix ltm/kcs commands broken by OVERRIDE_KERNEL="none"
Some ltm and kcs commands use parse_cli, but do not specify a kernel
because they are launching a ltm or kcs VM, or they are sending a
command to the ltm or kcs VM.  In that case OVERRIDE_KERNEL is set to
"none" and KERNEL will be set to the default kernel build directory.
If the latter does not exist, then get_kernel_file_info will fail, and
abort gce-xfstests.  Fix this by skipping the call to
get_kernel_file_info if OVERRIDE_KERNEL is set to "none".

Fixes: 966edd601810 ("run-fstests: get/infer the architecture from the kernel file")
Signed-off-by: Theodore Ts'o <tytso@mit.edu>