parse_cli: handle the case where the bzImage kernel is not in the build tree
The file(1) command can not distinguish between a 64-bit and a 32-bit
kernel.  So if the bzImage file is located in the build tree, we
disambiguate betweeh these two cases by examining the .config file in
the build tree.  However, if the bzImage has been copied outside of
the build tree, this trick won't work.  So in that case, if the
specified (or default) architecture is i386 or amd64, then assume that
the x86 kernel is also that architecture.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>