test-appliance: run shell script fragments after starting the LTM server
when starting the LTM server, run the shell script fragments in the
the directory gs://$GS_BUCKET/ltm-batch and gs://$GS_BUCKET/ltm-rc.
The shell script fragments in the ltm-batch directory are deleted
after they run, so they are executed one-time only.  This makes it
useful (for example) for regression testing.

The shell script fragments in the ltm-rc directory are run every time
that the LTM server is started.  This would be a good place to put
commands to set up watchers on a particular git repo, so after the ltm
server is stopped and restarted, the standard watchers will be restarted.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>