ltm: reboot hung tests and resume
Previously, a crashing or hung task aborted the rest of the test set.

Set the test kernel to reboot on panic to recover from crashes. Also,
allow LTM to force reboot a test VM if it has not updated it's status
in 1hr. This will help recover from hung tasks.

LTM pieces together the results from before and after the reboot. An
entry is inserted in the middle to mark the rebooted test's status as
"error" - using this status will distinguish these cases from regular
test failures.

Signed-off-by: Leah Rumancik <leah.rumancik@gmail.com>