gce-logger: force status metadata update on first test
I was running into an issue where early test crashes caused
noStatusTimeout to hit. gce-logger was sending the metadata update
comand at the start of the test but it was not completing before the
test crashed causing a reboot. To fix this and prevent false positives
with noStatusTimeout, wait for the status metadata update to finish
when running the first xfstest.

Signed-off-by: Leah Rumancik <leah.rumancik@gmail.com>