kvm-xfstests: allow specification of qemu machine type
For some reason, some large kernel configs generated by syzbot require
using the qemu machine type pc-q35-7.1.  The qemu machtype pc-q35-7.2,
which is aliased to the machine type q35 on the version of qemu
currently installed in Debian testing, will hang without showing any
kernel messages on the serial config.  The reason is unknown, but
allow the machine type to be specified by explicitly using the
--machtype command-line option.  The line "KVM_MACHTYPE=pc-q35-7.1"
can also be specified in $HOME/.config/kvm-xfstests if desired.

Link: https://github.com/google/syzkaller/issues/3740#issuecomment-1460996684
Signed-off-by: Theodore Ts'o <tytso@mit.edu>