gce-xfstests, test-appliance: disable the vm timeout when ltm runs tests
Add the --no-vm-timeout option to gce-xfststs which requests that the
test appliance disables the gce-finalize timer.  This timer is
responsible for forcibly cutting the test run short after 24 hours.

When test vm is managed by ltm, when a test takes more than an hour,
the VM will be forcibly restarted.  So we don't need the VM-level
timeout.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>