test-appliance: add support for continuing after a VM is preempted
We can determine whether the VM was preempted by querying the metadata
server in the 30 seconds grace period after VM is preempted.  If that
is the case hint that this is happened by setting the file
/results/preempted.  Then when the system is restarted, have
runtests.sh detect that this happened and restart the test was running
at the time the VM was preempted.

Also teach gce-xfstests the new subcommand simulate-maintenance-event.
This can be used to test a preemptible VM by using the command
"gce-xfstests simulate-maintenance-event <VM-NAME>".  Then after the
VM has stopped, then resume the VM using the command "gce-xfstests
start <VM-NAME>".

Signed-off-by: Theodore Ts'o <tytso@mit.edu>