gce-xfstests: add retry logic when starting VM's
Preemptible or spot VM's are more likely to fail due to lack of
resources in the GCE zone.  So add retry logic where we will retry
three times, and if enabled, fall back to using standard VM's if it's
not possible to start spot VM's.

Also transition to using spot VM's instead of preemptible VM's.  A
Spot VM is a newer version of preemptible VM's. and the primary
difference is that a Spot VM will not automtically terminate after 24
hours.  Also, --spot is a lot easier to type that --preemptible on the
gce-xfstests command-line.  :-)

Signed-off-by: Theodore Ts'o <tytso@mit.edu>