test-appliances: allow for an asynchronous reboot command
Systemd's reboot command is not instaneous and will return after
requesting systemd to reboot.  This can take a second or two, during
which the gce-load-kernel script will continue executing commands
intended for the kexec option.

Fix this by sleeping for a minute after the reboot command, and if the
system is still running force an immediate reboot using "reboot -f"

Signed-off-by: Theodore Ts'o <tytso@mit.edu>