gce-xfstests: teach gce-xfstests get-results the --no-cache option
If the user is only going to get a particular test result once,
there's no point having the compressed tar file take up space in /tmp.
So if the user uses the --no-cache option, delete the compressed tar
file containing the test results after displaying the requested
output.

If the --unpack or --unpack-dir options are given, assume that we
shouldn't cache the tar file, since everything the user might want
will be in the unpacked results directory.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>