test-appliance: add better debuggability for gce-load-kernel.log
Since the gce-load-kernel script often ends with a "reboot", nothing
ever gets written to the gce-load-kernel.log file.

Fix this by using "script -f", and while we're at it, move the check
whether gce-load-kernel has already populated the command line with
the fstestcfg parameter to kvm-xfstests.boot, so that we don't end up
calling the script on subsequent reboots.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>