test-appliance: use reboot -f in gce-load-kernel
In Debian Bookworm, for some reason, the "reboot" command as provided
by systemd doesn't seem to enter the BIOS boot scrippt, such that the
grub environment set by grub-reboot(8) is honored.  Instead, the first
(default) linux kernel is always reoobted, at least in a Google Cloud
VM.  The grub menu never gets displayed, and the grub-reboot setting
is not honored; and examining /boot/grub/grubenv after the reboot
command shows that it is not cleared.

The "reboot -f" command doesn't do a clean shutdown of the system, but
at least we go through the BIOS screen, and the grub menu does get
displayed, and the grub environment set by grub-reboot(8) is honored.

This is a regression relative to Debian Bullseye, where "reboot" would
do a clean shutdown, and then enter a reboot path where grub does
actually show the grub menu and grub-reboot is honored.  But for now,
this workarond is, while not ideal, is at least workable.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>