test-appliance: override PEP 668 for Debian Bookworm
Debian Bookworm started shipping Python 3.11 with support for PEP 668.
This disables the use of pip3 install for all but Python virtual
environments.  Unfortunately, doing things "right" with Python Virtual
environments would involve changing how we ship the xfstests-bld
specific Python scripts, which we currently ship via
test-appliances/file/usr/lib/python3/dist-packages.  This wouldn't be
that hard to make work with gce-xfstests, but we also need to make
things work for the test appliance for kvm-xfstests, which we are also
trying to keep lightweight.

So we'll hack around it for now.  Maybe at some point we'll do
something better, but this good enough for now.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>