test-appliance: use GCE metadata to figure out if VM is a ltm or kcs server
Previously, the gce-fetch-gs-files script would use the instance name
to determine whether the VM is the ltm or kcs server.  Unfortunately,
the gce-fetch-gs-file script is run out of systemd, to speed up the
startup time by running scripts in parallel.

Unfortuantely, despite the fact that gce-fs-fetch-files.service file
is not supposed to run until after google-network-setup.service is
completed, in the scripts provided with the Debian Bookworm, the
hostname is still set as "localhost", and has not yet been set to the
VM instance name.

So use "gce_attribute gce_xfs_{ltm,kcs}" to see if the metadata
indicating that the VM is supposed to be an LTM or KCS server instead
of keying off of the instance name.  This is arguably a better way to
do things, in case some day we want to be able support multiple (for
example) KCS servers.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>