test-appliance: run "make clean" if the compiler has changed
If the KCS repository cache contains object files or .*.cmd files
(which contain header file dependency information), these may be out
of date if the compiler has changed since the last time a kernel was
built in that directory.  This could be because the previous build was
using a different compiler, or the KCS has been upgraded to use a
newer compiler version.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>