test-appliance: run gce-shutdown out of kvm-xfstests.boot
Previously kvm-xfstests.boot would delete the gce-finalize-wait and
then allow the gce-finalize script to trigger gce-shutdown.  However,
with Debian Bookwork, somehow gce-shutdown would not complete before
the VM would shut itself down.  It's not entirely clear why that's the
case, but it's much simpler to just run gce-shutdown directly out of
kvm-xfstests.boot, so the kvm-xfstests.service systemd unit won't
complete until after the gce-shutdown has completed, and has started
the instance self-delete (which is asynchronous), so it will take a
few seconds to complete.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>