test-appliance: get the instance name from the metadata server
When the GCE scripts set the hostname to the GCE instance name is not
something that we can count upon; so just query the metadata server to
get the information and cache it in /var/lib/gce-xfstests/instance.
Between the fact that it can be a fully-qualified domain name, and set
at apparently random times while systemd unit files are getting run,
it's just safer to not rely on the hostname.

This fixes an issue where the test-run ID would sometimes get set to a
random value instead of one based on the VM instance name, based on
racy systemd unit files' run order on Debian Bookworm based test
appliances.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>