build-appliance: pass the date used for the tar timestamp to gen-image
Previously, gen-image would determine the timestamp to be used in the
tar file from the commit time of the head of the git repository.  The
problem is that gen-image can be running as root, and git now has
security checks which refuse it to operate on a repository owned by a
different user.  To avoid the "fatal: detected dubious ownership in
repository..." error, determine the timestamp that should be used in
the build-appliance script and pass it to the gen-image script via the
--src_date option.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>