build-appliance: create xfstests-<ARCH>.tar.gz
Create xfstests.<ARCH>.tar.gz in addition to the old xfstests.tar.gz
file.  We need this for doing release builds for kvm-xfstests.

In addition, when running "gce-xfstests --arch <ARCH>" look for
xfstests.<ARCH>.tar.gz so that we can be sure we create the image
using the correct set of test binaries.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>