gce-xfstests: make sure the /tmp/gce-xfstests.config.XXXXXX is deleted
The temp file /tmp/gce-xfstests.config.XXXXXX was cleaned up via a
relied on a "trap -.. EXIT" statement.  The problem is that the
parse_cli script also uses a trap statement to clean up files on exit.
Since we only need to use the temp file for a very short section of
the script, just delete the temp file once we're done needing it.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>