fstests-bld: use upstream libaio instead of outdated fork
fstests-bld is using a 12-year old in-tree fork of libaio that is
missing features such as support for RISC-V.  The only local changes are
arm64 support, which upstream has too, and autotools support which is
unnecessary (even for supporting CROSS_COMPILE).  Just use the upstream
git repo instead, at libaio-0.3.113 which is the latest version.

Note that libaio continues to be used as a static library, not shared.

Signed-off-by: Eric Biggers <ebiggers@google.com>