fstests-bld/get-all: replace directories that aren't git repos
When one of the target directories already exists but isn't actually a
git repo, the 'get-all' script gets confused and messes with the parent
repo "xfstests-bld" by changing the URL of the "origin" remote and
checking out a commit.  The upcoming change to use external libaio would
cause this issue to happen for anyone who has done a build and not
cleaned it.  Avoid this issue by making 'get-all' automatically replace
target directories that exist but aren't git repositories.

Signed-off-by: Eric Biggers <ebiggers@google.com>