kernel-configs: explicitly enable CONFIG_AUTOFS_FS for 6.5+ kernels
In 6.5, the last remnants of autofs v4 was removed.  In the config
files for 6.1+ kernels, CONFIG_AUTOFS4_FS implicitly enabled
CONFIG_AUTOFS_FS.  But when CONFIG_AUTOFS4_FS was removed,
CONFIG_AUTOFS_FS was not getting enabled by install-kconfig, and this
broke some systemd unit files which depends on automounting working.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>