qemu-xfstests: add a qemu-xfstests script as an alias for kvm-xfstests
Now that xfstests-bld is growing more and more support for non-x86
architectures (without using android-xfstests) the name "kvm-xfstests"
is getting a bit obsolete, since KVM is an x86-specific feature.  So
add the qemu-xfstests script as an acknowledgement to this fact.  We
still have the string "kvm" buried in far too many places, it's
probably not worth the pain of changing it, but we can at least
acknowledge this in the top-level user script.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>