setup-buildchroot: allow multiple fstests-bld dirs to share build chroots
If using multiple checkouts of fstests-bld stored in different
directories, it is possible to allow them to use the same build chroot.
To do this, if we are attempting to create a build chroot that already
exists, make sure the fstab includes a bind mount entry for the
fstests-bld from which we are calling setup-bulidchroot.

Signed-off-by: Leah Rumancik <leah.rumancik@gmail.com>