selftests: clean up how the helper function files are found by the tests
Allow selftests to be run without making assumptions about the current
directory.  The documentation currently states that they can be run
with the current directory being the selftests directory, in which
case they might be run as (for example) "./appliance" or in the
top-level directory, in which case they would be run as
"./selftests/appliance".

Signed-off-by: Theodore Ts'o <tytso@mit.edu>

The selftests can be run out of the top-level directory via
./selftests/<test-script>, e.g., ./selftests/appliance, or out of the
selftests directory, so they would be run via "./appliance".