test-appliance: change the LTM watcher to use a counter instead of a git id
Previously the test-id was based on the first 8 bytes of the git
commit id, which could potential be duplicated if the branch goes back
to git commit that has already been seen by the watcher.  So instead
make the test id to be based on a counter, and store the git commit in
a separate TestInfo field.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>