test-appliance: remove the LTM's watcher periodic repacking
The goal of repacking test results once a week into a single large tar
file was to reduce space (since the compression can be more
efficient).  The problem is that over time, the work needed to repack
and recompress the tar file gets larger and larger.  Since we hold the
history mutex during this process, it blocks the LTM server's
operations for far too long.

It also makes it harder to extract a particular test run's test
artifacts, so just remove it all.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>