test-appliance: kcs: resolve git branch name more reliably
If the user specifies a branch name to KCS, e.g., via "--repo origin"
we need to make sure that git repository is updated before trying to
resolve the branch name, since the reference in the locally cached
repository might be out of date.  On the other hand, the commit
reference is a SHA hash id, we'd rather not do a full git fetch if the
commit is already in the repository.

So the strategy we use is to try to resolve the commit first if it is
all hex digits and at least 7 characters long.  If that fails, then do
a git fetch, and then first try to resolve the commit prefixed with
"origin/", to see if it is branch name in the origin git tree.  If
that fails, then try resolving the bare commit name; this should work
if the commit revision specifier is either a tag or SHA hash id.

Also use the --prune (-p) and --force (-f) options so if the upstream
git branches are rewound or deleted, the local repository accurately
reflects the state of the upstream / remote repository.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>