Support using Linux kernel managed by git worktrees
Git worktrees use a .git file which contains a pointer for the
worktree's git directory (which is locasted as a subdirectory in the
common git directory under worktrees/<worktree_name>).  Account for
this in install-kconfig and kbuild scripts, as well as the
find_kernel_to_use helper function by kvm-xfstsets and gce-xfstests.

Tested via the following commands:
       kbuild --get-build-dir
       kbuild --get-kbuild-config
       install-kconfig
       kbuild --dpkg
       kvm-xfstests shell
       gce-xfstests upload-kernel

... with the following scenarios:
   1)  In a git tree witout any worktrees
   2)  In a worktree where there is a .git/kbuild/config file pointing
         the build tree at /build/ext4
   3)  In a worktree where there is a .git/kbuild/config file as above, and
         in addition, a worktree-specific kbuild config file in
        .git/worktrees/test/kbuild/config which points the build tree
         at /build/ext4-test.  (Where the worktree is named "test")

Signed-off-by: Theodore Ts'o <tytso@mit.edu>