run-tests/util: fix compatibility with openSUSE's file command
OpenSUSE has a non-standard magic data file for the 'file' command.
(It's gratuitiously different from the version of 'file' shipped with
Debian, Fedora, and MacOS.  Sigh...)  To work around this unfortunate
situation, include an excerpted version of the magic file from the
file version 5.45 and use it when trying to determine the architecture
of a compiled kernel.

Tested on Debian, Fedora, openSUSE, and MacOS.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>