test-appliance: drop local copies of attr and acl
For historical reasons, xfstests-bld had its own internal copies of
the attr and acl libraries and command-lione utilities.  This dates
back to 2011, when Debian didn't necessarily package all of the header
files needed for xfstests to be built.  In 2024, this is no longer
necessary, and the version in xfsteests-bld is older than what is
currently being shipped in Debian stable.

By dropping it, we save a bit space in the test appliance, and we
update to the latest version of the attr and acl packages.

NOTE: this will require installing the package libacl1-dev in any
existing build chroots.  The setup-buildchroot script has been updated
to install this package for any new build chroots, but existing build
chroots will need to manually install this package.  Alternatively,
xfststs-bld developers can just delete and recruit the buildchroot if
they prefer.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>