test-appliance: add support to use a mkfs config file for LTS kernels
Older LTS kernels don't necessarily support all of the file system
features of modern kernels.  So detect when we are testing an LTS
kernel, and if available, use an file system specific mkfs config file
(if available) to constrain the file system features needed for that
kernel version and file system type.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>