test-appliance: combine xfs's lts configuration with its mkfs option
mkfs.xfs has a massive mis-feature where it will not allow a
command-line option to be overridden by another command-line option.
Worse, if there is a config file which defines the file system
configuration used three years ago, if there is a command-line which
overrides that option, it will fail.  This makes testing XFS for LTS
kernels for 5.15.... painful.

To work around this, we will merge the LTS-X.YY config file with the
mkfs options used for a particular test configuration, doing something
in shell script functions to work around mkfs.xfs's misdesign.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>