gce-create-image: add support for building the image using Debian Trixie
Debian 13 (Trixie) is the current testing version of the Debian, with
a hard freeze scheduled for mid-May 2025, with a final release
probably sometime in late Summer 2025.  The GCE packages for Trixie
haven't been published yet, so the GCE image is built by starting with
a Debian 12 (Bookworm) image and then upgrading it using apt-get
dist-upgrade.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>