test-appliance: drop a exclusion of a local test for an io_uring bug
We had a local test which tested a bug in io_uring which would old a
file system mount busy causing umounts to fail.  The test had gotten
renumbered a number of times, and we forgot to edit the exclusion.  It
appears the test is not failing, so let's just drop the incorrect
exclusiosn.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>